Item generation error implemented
FailedItemGen will be raised when failed to generate a specific item with more than 10 trials. Error will informa the item's name, rarity, valid names, exteriors and the converted item name.

While this might be productive for development it would probably be better if the error handler cog also catches it and informs the user.

Also moved ModelExtMixin to new file for development